pyproject.toml
[project]
name = "trackmate-mcp"
version = "0.1.0"
description = "Smart delivery tracking assistant MCP - AI-powered package tracking with status translation and problem diagnosis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.2.0",
    "fastmcp>=0.4.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
trackmate = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
